refactor: move asset parsing to another crate
