+++ Cargo.toml
@@ -3,6 +3,7 @@ members = [
     "game",
     "crates/bevy_renderware",
     "crates/renderware-format",
+    "crates/vice-city-formats",
     "tools/renderware-format-converter",
     "tools/renderware-format-dumper",
 ]
