+++ crates/vice-city-formats/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "vice-city-formats"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+bitflags = "1.3.2"
+glam = "0.20.5"
