+++ crates/vice-city-formats/src/common.rs
