+++ crates/vice-city-formats/src/lib.rs
@@ -1,0 +1,8 @@
+mod common;
+
+pub mod dat;
+pub mod ide;
+pub mod ipl;
+
+pub use ide::Ide;
+pub use ipl::Ipl;
